gvpr: use 'size_t' variables for counting blocks
Squashes a -Wsign-conversion compiler warning.

+++ lib/gvpr/compile.h
@@ -73,7 +73,7 @@ extern "C" {
 	int flags;
 	Expr_t *prog;
 	Exnode_t *begin_stmt;
-	int n_blocks;
+	size_t n_blocks;
 	comp_block  *blocks;
 	Exnode_t *endg_stmt;
 	Exnode_t *end_stmt;
